requirements.txt
absl-py==0.4.1
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
Cython==0.29
decorator==4.3.0
enum34==1.1.6
html5lib==1.0.1
ipython==5.8.0
ipython-genutils==0.2.0
numpy==1.15.1
pandas==0.23.4
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.4
plyvel==1.0.5
prompt-toolkit==1.0.15
protobuf==3.6.1
ptyprocess==0.6.0
Pygments==2.2.0
python-dateutil==2.7.3
pytz==2018.5
re3==0.2.23
scandir==1.9.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
